Made code Python 2 compatible. #1
